+++ src/init/discovery.clj
@@ -2,7 +2,18 @@
   (:require [clojure.string :as str]
             [init.config :as config]
             [init.component :as component]
-            [init.meta :as meta]))
+            [init.meta :as meta])
+  (:import [java.net JarURLConnection URL]
+           [java.nio.file
+            FileSystem
+            FileSystems
+            FileVisitResult
+            FileVisitor
+            Files
+            Path
+            Paths]))
+
+(set! *warn-on-reflection* true)
 
 (defn from-namespaces
   "Builds a configuration from the given namespaces."
@@ -51,24 +62,87 @@
 
 ;;;; CLASSPATH SCANNING
 
+(defn- filename [^Path p]
+  (.. p getFileName toString))
+
+(defn- namespace-unmunge [n]
+  (.replace (str n) \_ \-))
+
+(defn- scan-dir [^Path start]
+  (let [names (volatile! ())
+        found (volatile! ())]
+    (Files/walkFileTree
+     start
+     (reify FileVisitor
+       (preVisitDirectory [_ dir _attrs]
+         (let [n (filename dir)]
+           (if (re-matches #"\w+" n)
+             (do
+               (when (not= start dir)
+                 (vswap! names conj (namespace-unmunge n)))
+               FileVisitResult/CONTINUE)
+             FileVisitResult/SKIP_SUBTREE)))
+       (postVisitDirectory [_ dir _exc]
+         (when (not= start dir)
+           (vswap! names pop))
+         FileVisitResult/CONTINUE)
+       (visitFile [_ path _attrs]
+         (when-let [[_ n] (re-matches #"(\w+)(__init\.class|\.cljc?)" (filename path))]
+           (->> (conj @names (namespace-unmunge n))
+                reverse
+                (str/join ".")
+                (vswap! found conj)))
+         FileVisitResult/CONTINUE)
+       (visitFileFailed [_ path exc]
+         (throw (ex-info (str "Failed to read " path) {:path path} exc)))))
+    @found))
+
+(defn- split-jar-url [^URL url]
+  (let [^JarURLConnection conn (.openConnection url)]
+    [(Paths/get (.. conn getJarFileURL toURI)) (.getEntryName conn)]))
+
+(defn- jar-file-system ^FileSystem [^Path path]
+  (let [^ClassLoader cl nil]
+    (FileSystems/newFileSystem path cl)))
+
+(defn- scan-jar [url]
+  (let [[jar entry] (split-jar-url url)]
+    (with-open [fs (jar-file-system jar)]
+      (scan-dir (.getPath fs entry (into-array String nil))))))
+
+(defn- scan-url [^URL url]
+  (case (.getProtocol url)
+    "file" (scan-dir (Paths/get (.toURI url)))
+    "jar"  (scan-jar url)))
+
+(defn- scan-classpath
+  ([^ClassLoader cl]
+   (->> (.getResources cl "")
+        enumeration-seq
+        (mapcat scan-url)
+        (map symbol)))
+  ([^ClassLoader cl prefix]
+   (->> (str/replace (namespace-munge prefix) \. \/)
+        (.getResources cl)
+        enumeration-seq
+        (mapcat scan-url)
+        (map #(symbol (str prefix "." %))))))
+
+(defn- context-classloader []
+  (.. Thread currentThread getContextClassLoader))
+
 (defn classpath-namespaces
   "Returns a set of symbols of all namespaces on the classpath.  When
    `prefixes` are given, only returns namespaces starting with one of the
-   prefixes.
-
-   Requires `org.clojure/java.classpath` and `org.clojure/tools.namespace`
-   on the classpath."
+   prefixes."
   ([]
-   (let [classpath       (requiring-resolve 'clojure.java.classpath/classpath)
-         find-namespaces (requiring-resolve 'clojure.tools.namespace.find/find-namespaces)]
-     (set (find-namespaces (classpath)))))
+   (into #{} (scan-classpath (context-classloader))))
   ([prefixes]
-   (filter (some-prefix-pred prefixes) (classpath-namespaces))))
+   (let [cl (context-classloader)]
+     (into #{} (mapcat (partial scan-classpath cl)) prefixes))))
 
 (defn load-namespaces
-  "Loads and returns all namespaces matching `prefixes` on the classpath.
-
-   See [[classpath-namespaces]] for required libraries."
+  "Loads and returns all namespaces matching `prefixes` on the classpath."
   [prefixes]
   (let [ns-names (classpath-namespaces prefixes)]
     (run! require ns-names)
@@ -76,37 +150,26 @@
 
 (defn scan
   "Scans the classpath for namespaces starting with `prefixes`, and returns
-   a config with all components found in those.
-
-   See [[classpath-namespaces]] for required libraries."
+   a config with all components found in those."
   ([prefixes]
    (from-namespaces (load-namespaces prefixes)))
   ([config prefixes]
    (from-namespaces config (load-namespaces prefixes))))
 
-(defn- emit-config [config]
-  (let [libspecs      (->> (vals config)
-                           (map #(-> %1 :var meta :ns ns-name))
-                           (into #{})
-                           (map #(list `quote %)))
-        static-config (update-vals config (fn [{:keys [var hook-vars]}]
-                                            `(meta/component ~var ~hook-vars)))]
-    `(do (require ~@libspecs)
-         ~static-config)))
+(defn- emit-static-scan [prefixes]
+  (let [namespaces (classpath-namespaces prefixes)]
+    `(do
+       (require ~@(map (partial list 'quote) namespaces))
+       (discovery/from-namespaces
+        ~(mapv (fn [n] `(the-ns (quote ~n))) namespaces)))))
 
 (defmacro static-scan
-  "Like [[scan]], but discovers the configuration at macro expansion time.
-   As such, libraries required for [[classpath-namespaces]] are only needed
-   during compilation, not at runtime.
+  "Like [[scan]], but scans the classpath at macro expansion time.
 
    Expands to code that uses `require` to load required namespaces, and a
-   config map creating var components with [[init.meta/component]].
-
-   Designed to be used in ahead-of-time compiled applications, that have
-   `clojure.tools.namespace` as a development dependency but don't want to
-   have it as a runtime dependency."
+   config map created with [[from-namespaces]]."
   [prefixes]
-  (-> (eval prefixes) scan emit-config))
+  (emit-static-scan (eval prefixes)))
 
 ;;;; SERVICE REGISTRY