+++ test/init/discovery_test.clj
@@ -9,3 +9,13 @@
     (is (pred 'init.discovery-test.deeply.nested.child))
     (is (not (pred 'init)))
     (is (not (pred 'init.discovery-test-sibling)))))
+
+(deftest classpath-namespaces-test
+  (let [namespaces (discovery/classpath-namespaces)]
+    (is (contains? namespaces (ns-name *ns*)))
+    (is (contains? namespaces 'init.core)))
+  (let [namespaces (discovery/classpath-namespaces ['init])]
+    (is (contains? namespaces (ns-name *ns*)))
+    (is (contains? namespaces 'init.core)))
+  (let [namespaces (discovery/classpath-namespaces ['init.core 'init.foobar])]
+    (is (empty? namespaces))))